+++ software_esp32/include/VdmConfig.h
@@ -93,6 +93,18 @@ typedef struct  {
   uint8_t publishPlainText : 1;
 } VDM_PROTOCOL_CONFIG_FLAGS;
 
+typedef struct  {
+  uint8_t timeoutActive : 1 ;
+} VDM_PROTOCOL_MQTT_CONFIG_FLAGS;
+
+
+typedef struct  {
+  VDM_PROTOCOL_MQTT_CONFIG_FLAGS flags;
+  uint32_t timeOut;
+  uint8_t toPos;
+} VDM_PROTOCOL_MQTT_CONFIG;
+
+
 typedef struct {
   uint8_t  dataProtocol; // 0 = no protocol , 1 = mqtt
   uint32_t brokerIp;
@@ -104,7 +116,10 @@ typedef struct {
   VDM_PROTOCOL_CONFIG_FLAGS protocolFlags;
   uint16_t keepAliveTime;
   uint32_t minBrokerDelay;
+  VDM_PROTOCOL_MQTT_CONFIG mqttConfig;
 } VDM_PROTOCOL_CONFIG;
+
+
 
 typedef struct {
   char  name[11];
@@ -167,6 +182,8 @@ typedef struct  {
   uint8_t notDetect : 1;
   uint8_t mqttTimeOut : 1;
   uint8_t reset : 1;
+  uint8_t ds18Failed : 1;
+  uint8_t tValueFailed : 1;
 } VDM_MSG_REASON_CONFIG_FLAGS;
 
 typedef struct  {
@@ -247,6 +264,9 @@ typedef struct {
 #define nvsProtBrokerPublishFlags   "brokerPF"
 #define nvsProtBrokerKeepAliveTime  "brokerKAT"
 #define nvsProtBrokerMinBrokerDelay "brokerMD"
+#define nvsProtBrokerMQTTFlags      "brokerMQF"
+#define nvsProtBrokerMQTTTimeOut    "brokerMQTO"
+#define nvsProtBrokerMQTTToPos      "brokerMQToPos"
 
 #define nvsValvesCfg                "valvesCfg"
 #define nvsValvesControlCfg         "valvesCtrlCfg"